move check_dummy_inputs_allowed to common export utils
